Move distributed test conftest
